pyproject.toml
[project]
name = "trainy"
version = "0.1.0"
description = "Training tracking web application with FIT file import and AI planning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "fitparse>=1.2.0",
    "httpx>=0.28.1",
    "instructor>=1.14.4",
    "jinja2>=3.1.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.17",
    "sse-starlette>=2.1.0",
    "uvicorn[standard]>=0.32.0",
]

[project.scripts]
trainy-import = "scripts.import_activities:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
